Move NtUtil.NtToRubyTime from gems/pending/fs/ntfs to gems/pending/util/win32
