+++ gems/pending/fs/ntfs/NtfsAttribFileName.rb
@@ -1,4 +1,5 @@
 require 'fs/ntfs/NtUtil'
+require 'util/win32/nt_util'
 
 require 'binary_struct'
 require 'util/miq-unicode'
