+++ gems/pending/fs/ntfs/NtfsAttribStandardInformation.rb
@@ -1,4 +1,4 @@
-require 'fs/ntfs/NtUtil'
+require 'util/win32/nt_util'
 
 require 'binary_struct'
